requirements.txt
pygithub==2.1.1
alist-sdk